Move Customs Office admin endpoint onto available routes in preperation for the Admin being made available in Production
